chore-api: 애플리케이션 패키지 이름 변경
- 도메인을 확실하기 분리할 수 있도록 아래와 같이 패키지를 변경
    - `plan`을 `travel_plan_comment`로 변경
    - `travel`에 있던 `TravelPlan` 관련 코드를 `travel_plan`으로 이동
