+++ server/src/main/java/com/yeohaeng_ttukttak/server/application/travel_plan/dto/UpdateTravelPlanRequest.java
@@ -1,4 +1,4 @@
-package com.yeohaeng_ttukttak.server.application.travel.controller.dto;
+package com.yeohaeng_ttukttak.server.application.travel_plan.dto;
 
 import com.yeohaeng_ttukttak.server.application.travel.service.dto.UpdateTravelVisitCommand;
