+++ server/src/main/java/com/yeohaeng_ttukttak/server/application/travel_plan_comment/dto/WriteCommentRequest.java
@@ -1,7 +1,6 @@
-package com.yeohaeng_ttukttak.server.application.plan.dto;
+package com.yeohaeng_ttukttak.server.application.travel_plan_comment.dto;
 
 import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
 
 public record WriteCommentRequest(
         @NotEmpty String content
